chore: Change admin path for Building controller
